CMake: Fix conditional compilation of PUML tests
When `CMAKE_CXX_STANDARD` is not set the LESS condition is not satisfied
and the puml tests are still added even if the compiler doesn't support
it or it isn't enabled.

Check for it being set first.
Otherwise add those tests as a separate target if the compiler does support C++20.